Rest_and_Forms/requirements.txt
Django==1.10.4
django-solo==1.1.2
django-wysiwyg-redactor==0.4.9.1
djangorestframework==3.5.3
